+++ libstuff/SLog.cpp
@@ -7,7 +7,7 @@ atomic<int> _g_SLogMask(LOG_INFO);
 void SLogStackTrace() {
     // Output the symbols to the log
     void* callstack[100];
-    int depth = backtrace(callstack, 100);
+    int64_t depth = backtrace(callstack, 100);
     vector<string> stack = SGetCallstack(depth, callstack);
     for (const auto& frame : stack) {
         SWARN(frame);
